+++ src/app/sign-in/reducer/SignInPageReducer.ts
@@ -1,29 +1,5 @@
-export type SignInPageAction =
-  | { type: 'INPUT_EMAIL'; value: string }
-  | { type: 'INPUT_PASSWORD'; value: string }
-  | { type: 'VALIDATE_EMAIL'; error: string }
-  | { type: 'VALIDATE_PASSWORD'; error: string }
-  | { type: 'REQUEST_SIGN_IN' }
-  | { type: 'SUCCESS_SIGN_IN' }
-  | { type: 'FAILURE_SIGN_IN'; error: string };
-
-export type SignInPageState = {
-  email: string;
-  password: string;
-  emailValidateErrorMessage: string;
-  passwordValidateErrorMessage: string;
-  isRequesting: boolean;
-  signInErrorMessage: string;
-};
-
-export const initialSignInPageState: SignInPageState = {
-  email: '',
-  password: '',
-  emailValidateErrorMessage: '',
-  passwordValidateErrorMessage: '',
-  isRequesting: false,
-  signInErrorMessage: '',
-};
+import { SignInPageAction } from '@/app/sign-in/action/SignInPageAction';
+import { SignInPageState } from '@/app/sign-in/state/SignInPageState';
 
 export function signInPageReducer(
   state: SignInPageState,
